reorder types
